// src/pages/App/Admin/Abilities/Create/validation/schema.ts
import { z } from 'zod';

import { requiredError } from '@/utils/validation/errors';

export const AbilitiesSchema = z.object({
  categoryType: z.string().nullable().optional(),
  category: z.string(requiredError).trim().min(2, { message: 'Digite um nome válido' }),
  ability: z.string(requiredError).trim().min(2, { message: 'Digite um assunto válido' }),
  source: z.string(requiredError).trim().min(2, { message: 'Digite uma origem válida' }),
});

export type AbilitiesSchemaType = z.infer<typeof AbilitiesSchema>;